Ref issue #98 Add solvcon.parcel.gas.oblique_shock for calculation of oblique shock wave relations.
Based on *Modern Compressible Flow: With Historical Perspective* 3/e by J. D. Anderson, organize the relevant oblique shock wave relations that can be used for calculating the flow properties in all the three zones of the problem of oblique shock reflection from wall.

Full description of the reflection problem and the formulation is yet to be added.