Ref issue #122 Manage lookup tables in Block class with descriptor
Add solvcon.block._TableDescriptor to manage the lookup table objects (of both Table and ndarray types) in solvcon.block.Block.  In Block, attributes of the following names will be managed with _TableDescriptor:

	ndcrd, fccnd, fcnml, fcara, clcnd, clvol, fctpn, cltpn, clgrp, fcnds, fccls, clnds, clfcs

In addition to the names alone, other attributes derived from those names are under the management.  They are prefixed with 'tb', 'sh', and 'gst'.  The unprefixed names are for body arrays.  Each prefix denotes a table group.  There are 4 groups in total.  Tables of each group are contained in a designated dict:

	- _table for prefix 'tb'
	- _shared_array for prefix 'sh'
	- _body_array for prefix '' (no prefix)
	- _ghost_array for prefix ''gst'

To assign the many descriptors, the meta class BlockMeta is assigned to Block.  The above logic is clearly implemented in BlockMeta.__new__().

In the future I want to make solvcon.mesh.Mesh manages the lifecycle of those tables, and solvcon.block.Block bypasses requests of the tables to Mesh.  I think these descriptors will help the plan.